add e2e sample test
