+++ apps/mudita-center-e2e/src/specs/overview/overview-menu-item.e2e.ts
@@ -4,12 +4,12 @@ describe("Overview Menu Item", () => {
   it("Overview menu item is displayed", async () => {
     await browser.waitUntil(() => E2EMockClient.checkConnection())
 
-    E2EMockClient.addDevice()
+    E2EMockClient.addDevice({ path: "123123", serialNumber: "asdf" })
 
     const menuItem = await $(`//a[@href="#/generic/mc-overview"]`)
 
     expect(menuItem).toBeDisabled()
 
-    await browser.pause(100000)
+    // await browser.pause(100000)
   })
 })
