+++ libs/e2e-mock/server/src/lib/mock-device/mock-device.ts
@@ -4,7 +4,7 @@
  */
 
 import { Result, ResultObject } from "Core/core/builder"
-import { DeviceType, ResponseStatus } from "Core/device"
+import { DeviceType } from "Core/device"
 import { BaseDevice } from "Core/device/modules/base-device"
 import { ApiResponse } from "Core/device/types/mudita-os"
 import {
@@ -13,187 +13,7 @@ import {
   APIRequestWithPayload,
 } from "device/models"
 import { PortInfo } from "serialport"
-
-type MethodObject = Partial<
-  Record<
-    APIMethodsType,
-    { every?: ApiResponse<unknown>; once?: ApiResponse<unknown> }
-  >
->
-
-const DEFAULT_RESPONSES: Partial<Record<APIEndpointType, MethodObject>> = {
-  API_CONFIGURATION: {
-    GET: {
-      every: {
-        status: ResponseStatus.Ok,
-        body: {
-          apiVersion: "1.0.0",
-          lang: "en-US",
-          variant: "black",
-          features: ["mc-overview"],
-          productId: "2006",
-          vendorId: "0e8d",
-          serialNumber: "0123456789ABCDEF",
-        },
-      },
-    },
-  },
-  MENU_CONFIGURATION: {
-    GET: {
-      every: {
-        status: ResponseStatus.Ok,
-        body: {
-          title: "Kompakt",
-          menuItems: [
-            {
-              feature: "mc-overview",
-              displayName: "Overview",
-              icon: "overview",
-            },
-          ],
-        },
-      },
-    },
-  },
-  OUTBOX: {
-    GET: {
-      every: {
-        status: ResponseStatus.Ok,
-        body: { features: [], data: [] },
-      },
-    },
-  },
-  FEATURE_DATA: {
-    GET: {
-      every: {
-        status: ResponseStatus.Ok,
-        body: {
-          summary: {
-            about: {
-              serialNumber: { text: "0123456789ABCDEF" },
-              imei1: { text: "864055030138811" },
-              imei2: { text: "864055030138829" },
-              sar: {
-                text: "### SAR\n\nLorem ipsum dolor sit amet, consectetur adipiscing elit. Donec sed aliquet ligula, viverra feugiat massa. In hac habitasse platea dictumst.\n\n1. Interdum et malesuada fames ac ante ipsum primis in faucibus.\n2. Suspendisse consectetur, nibh non consequat hendrerit, nibh felis commodo lacus, id auctor ante purus vitae justo.\n3. Cras purus neque, pharetra vitae nulla ac, mollis facilisis felis. Sed sit amet ex diam.\n\n> Sed accumsan sem nec iaculis euismod.",
-              },
-            },
-          },
-          sections: {
-            battery: { icon: "battery-charging-5", text: "100%", subText: "" },
-            update: { text: "ANDROID 12", version: "0.3.0" },
-            status: { badgeText: "Offline+" },
-            "airplane-mode": { icon: "airplane-mode", text: "Airplane mode" },
-          },
-        },
-      },
-    },
-  },
-  FEATURE_CONFIGURATION: {
-    GET: {
-      every: {
-        status: ResponseStatus.Ok,
-        body: {
-          title: "Overview",
-          summary: {
-            show: true,
-            showImg: true,
-            imgVariant: "black",
-            showSerialNumber: true,
-            serialNumberLabel: "Serial number",
-            showAbout: true,
-            aboutTitle: "About your device",
-            aboutIcon: "device",
-            aboutSubtitle: "Device details",
-            aboutFields: [
-              {
-                dataKey: "serialNumber",
-                type: "detail-list-text",
-                title: "Serial number",
-              },
-              {
-                dataKey: "imei1",
-                type: "detail-list-text",
-                title: "IMEI (sim slot 1)",
-              },
-              {
-                dataKey: "imei2",
-                type: "detail-list-text",
-                title: "IMEI (sim slot 2)",
-              },
-              {
-                dataKey: "sar",
-                type: "detail-list-modal",
-                title: "SAR",
-                buttonText: "Check SAR information",
-              },
-            ],
-          },
-          sections: [
-            {
-              title: "Status",
-              type: "tile-list",
-              dataKey: "status",
-              fields: [
-                { dataKey: "battery", type: "icon-text" },
-                { dataKey: "airplane-mode", type: "icon-text" },
-              ],
-            },
-            {
-              dataKey: "update",
-              type: "mc-overview-update",
-              title: "Android OS",
-              currentVersionKey: "version",
-              showBadge: false,
-              versionLabel: "Current version:",
-            },
-            {
-              dataKey: "backup",
-              type: "mc-overview-backup",
-              title: "Backup",
-              backupFeatures: [
-                { label: "Contacts list", key: "CONTACTS_LIST" },
-                { label: "Call log", key: "CALL_LOG" },
-                { label: "Messages", key: "MESSAGES" },
-                { label: "Notes", key: "NOTES" },
-                { label: "Calendar events", key: "CALENDAR_EVENTS" },
-                {
-                  label: "OS version & OS settings",
-                  key: "OS_VERSION_AND_SETTINGS",
-                },
-                { label: "App settings: Phone, Messages", key: "APP_SETTINGS" },
-              ],
-              restoreFeatures: [
-                {
-                  label: "Contacts list",
-                  feature: "CONTACTS_LIST",
-                  keys: ["CONTACTS_LIST"],
-                },
-                { label: "Call log", feature: "CALL_LOG", keys: ["CALL_LOG"] },
-                { label: "Messages", feature: "MESSAGES", keys: ["MESSAGES"] },
-                { label: "Notes", feature: "NOTES", keys: ["NOTES"] },
-                {
-                  label: "Calendar events",
-                  feature: "CALENDAR_EVENTS",
-                  keys: ["CALENDAR_EVENTS"],
-                },
-                {
-                  label: "OS version & OS settings",
-                  feature: "OS_VERSION_AND_SETTINGS",
-                  keys: ["OS_VERSION_AND_SETTINGS"],
-                },
-                {
-                  label: "App settings: Phone, Messages",
-                  feature: "APP_SETTINGS",
-                  keys: ["APP_SETTINGS"],
-                },
-              ],
-            },
-          ],
-        },
-      },
-    },
-  },
-}
+import DEFAULT_RESPONSES from "./default-responses"
 
 export class MockDevice extends BaseDevice {
   public connect(): Promise<ResultObject<undefined>> {
@@ -205,17 +25,14 @@ export class MockDevice extends BaseDevice {
   public request<R, T extends APIEndpointType>(
     config: APIRequestWithPayload<T>
   ): Promise<unknown> {
-    const y = config.method
     const responses =
       DEFAULT_RESPONSES[config.endpoint]?.[config.method as APIMethodsType]
 
-    // console.log(responses)
-
-    if (responses?.every) {
+    if (responses) {
       return new Promise<ResultObject<ApiResponse<unknown>>>((resolve) => {
         resolve(
           Result.success({
-            ...responses.every,
+            ...responses,
             endpoint: config.endpoint,
           } as unknown as ApiResponse<unknown>)
         )
